backend/requirements.txt
easyocr==1.7.2
filelock==3.20.0
fsspec==2025.12.0
ImageIO==2.37.2
Jinja2==3.1.6
lazy_loader==0.4
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
ninja==1.13.0
numpy==2.2.6
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==12.0.0
pyclipper==1.4.0
python-bidi==0.6.7
PyYAML==6.0.3
scikit-image==0.25.2
scipy==1.16.3
setuptools==80.9.0
shapely==2.1.2
sympy==1.14.0
tifffile==2025.12.12
torch==2.9.1
torchvision==0.24.1
typing_extensions==4.15.0
Flask==2.3.2
requests==2.32.0
